4 enhancements / bug fixes:
Added a global variable gamemode to enable a 1-player and 2-player mode. It is simply hardcoded for now, but can of course be made user-selectable if desired.

The paddles could disappear from the screen. Constrained the movement of the paddles to the top and bottom of the playing field.

The ball would stick to the paddle if it touched the paddle near the edges. Solved in the code section under reset_ball.

Implemented the function computer_move() to add an 'AI'-player.